// client/app/api/accounting/model/BusinessCategoryModel.ts
/**
 * GuanPlus.AccountingFirm.WebApi
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import * as models from './models';

export interface BusinessCategoryModel {
    /**
     * id
     */
    id?: string;

    /**
     * 名称
     */
    name?: string;

    /**
     * 级别
     */
    level?: number;

    /**
     * 类别：收入或者支出
     */
    businessCategoryType?: models.BusinessCategoryTypeEnumModel;

    /**
     * 禁止创建子级
     */
    lock?: boolean;

    /**
     * 父级交易类别
     */
    parentBusinessCategory?: models.IdNameModel;

    /**
     * 顺序
     */
    order?: number;

    /**
     * 是否是初始化
     */
    isInit?: boolean;

    /**
     * 是否显示
     */
    display?: boolean;

}